Make CommandRunner return an error if command fails
Indeed, commands can take too long to terminate or file does not exist
or command is not executable.

+++ src/CommandRunner.cpp
@@ -1,16 +1,21 @@
 #include "CommandRunner.hpp"
-#include <memory>
-#include <stdio.h>
+
+#include <errno.h>
 #include <fstream>
-#include <sstream>
 #include <iostream>
+#include <memory>
 #include <signal.h>
+#include <sstream>
+#include <stdio.h>
+#include <sys/stat.h>
 #include <sys/types.h>
 #include <sys/wait.h>
 #include <time.h>
-#include <errno.h>
+#include <unistd.h>
 
 #include <glog/logging.h>
+
+#include <stout/nothing.hpp>
 #include <stout/try.hpp>
 
 #define TEMP_FILE_TEMPLATE "/tmp/criteo-mesos-XXXXXX"
@@ -70,6 +75,15 @@ public:
 private:
   std::string m_filepath;
 };
+
+inline bool fileExists(const std::string& name) {
+  struct stat buffer;   
+  return (stat (name.c_str(), &buffer) == 0); 
+}
+
+inline bool isFileExecutable(const std::string& name) {
+ return !access(name.c_str(), X_OK);
+}
 
 
 /*
@@ -137,11 +151,13 @@ pid_t popen2(const std::string& command, const std::vector<std::string>& args,
  * Fork the process to run command and kill the child if it does not
  * finish before the timeout deadline.
  *
+ * TODO(clems4ever): split this method so that it becomes easier to read.
+ *
  * @param command The command to execute in the child process.
  * @param timeout The timeout deadline in seconds before killing the
  * child process.
  */
-void runCommandWithTimeout(const std::string& command,
+Try<Nothing> runCommandWithTimeout(const std::string& command,
   const std::vector<std::string>& args, int timeout) {
   int status;
   unsigned int tick = 0;
@@ -151,12 +167,23 @@ void runCommandWithTimeout(const std::string& command,
   const unsigned long long terminationTicks = SECONDS / TEN_MS;
   unsigned long long totalTicks = processTicks + terminationTicks;
   struct timespec timeoutSpec = {0, TEN_MS};
-  pid_t pid = popen2(command, args);
   bool forceKill = false;
+  bool hasError = false;
+
+  if(!fileExists(command)) {
+    return Error("No such file or directory: \"" + command + "\"");
+  }
+
+  if(!isFileExecutable(command)) {
+    return Error("File is not executable: \"" + command + "\"");
+  }
+
+  pid_t pid = popen2(command, args);
 
   while (tick < totalTicks) {
     nanosleep(&timeoutSpec, NULL);
     if (tick == processTicks) {
+      hasError = true;
       LOG(WARNING) << "External command took too long to exit. "
         << "Sending SIGTERM...";
       if(kill(pid, SIGTERM) == -1) {
@@ -169,6 +196,7 @@ void runCommandWithTimeout(const std::string& command,
     if (rc < 0) {
       LOG(ERROR) << "Error when waiting for child process running the "
                  << "external command: " << strerror(errno);
+      hasError = true;
       forceKill = true;
       break;
     }
@@ -181,9 +209,18 @@ void runCommandWithTimeout(const std::string& command,
   if(forceKill || tick == totalTicks) {
     LOG(WARNING) << "External command is still running. Sending SIGKILL...";
     if(kill(pid, SIGKILL) == -1) {
+      hasError = true;
       LOG(ERROR) << "Failed to kill the command: " << strerror(errno);
     }
+    else {
+      return Error("Command \"" + command + "\" was too long to return");
+    }
   }
+
+  if(hasError) {
+    return Error("Failed to successfully run the command \"" + command + "\"");
+  }
+  return Nothing();
 }
 
 /*
@@ -199,7 +236,7 @@ void runCommandWithTimeout(const std::string& command,
  * @param timeout The timeout deadline before killing the child process.
  * @return The output of the command read from the output file.
  */
-std::string run(const std::string& command, const std::string& input,
+Try<std::string> run(const std::string& command, const std::string& input,
   int timeout) {
   try {
     TemporaryFile inputFile;
@@ -212,12 +249,14 @@ std::string run(const std::string& command, const std::string& input,
     std::vector<std::string> args;
     args.push_back(inputFile.filepath());
     args.push_back(outputFile.filepath());
-    runCommandWithTimeout(command, args, timeout);
+    Try<Nothing> result = runCommandWithTimeout(command, args, timeout);
+    if(result.isError()) {
+      return Error(result.error());
+    }
     return outputFile.readAll();
   }
   catch(const std::runtime_error& e) {
-    LOG(ERROR) << e.what();
-    return std::string();
+    return Error("Unable to run \"" + command + "\": " + e.what());
   }
 }
